+++ CHANGELOG.md
@@ -6,4 +6,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Implement quasi-binary decoder.
+- Implement Serial, a node-serialport wrapper that allows path changes and keeps track of port state.
 - Implement Parser, a general purpose parser meant to be implemented and fitted to anyone's needs.